Add files via upload
My pharmacy simulator. Get a first person perspective into the responsibilities of a pharmacy employee. The program features the ability to purchase items from Kinray [a medicine wholesaler], the ability to read and write electronic scripts, and work with role-based access controls. Be careful - impersonation is legal in the United States, and very much legal within the realm of this program.