Add files via upload
The Love-O-Meter is a project featuring the use of a temperature sensor to detect heat. Using a set of formulas, the Arduino will print the temperature once every pre-determined instance. By touching the temperature sensor, it is possible to elevate the temperature of the sensor, which will inherently light up the corresponding LEDs based on the difference between current temperature and the baseline temperature.